Support dynamic loading of JavaScript commands
