+++ src/commands/command_module_loader.ts
@@ -1,10 +1,11 @@
 import { CommandModuleCache } from './command_module_cache';
 import { IShellWorker } from '../defs_internal';
-import { default as EmscriptenModuleFactory } from '../fs';
+import { default as EmscriptenModuleFactory } from '../types/fs';
+import { JavaScriptModule } from '../types/javascript_module';
 
 /**
- * Loader of JavaScript/WASM modules. Once loaded, a module is cached so that it is faster to
- * subsequently.
+ * Loader of JavaScript/WebAssembly modules. Once loaded, a module is cached so that it is faster
+ * to subsequently.
  * Must be run in a WebWorker.
  */
 export class CommandModuleLoader {
@@ -14,25 +15,43 @@ export class CommandModuleLoader {
   ) {}
 
   /**
-   * Load the specified WebAssembly module.
+   * Load the specified JavaScript command module.
    * If loading fails return undefined, and caller is responsible for reporting this to the user if
    * appropriate.
    */
-  public getModule(
+  public getJavaScriptModule(
+    packageName: string,
+    moduleName: string
+  ): JavaScriptModule | undefined {
+    const module = this._getModule(packageName, moduleName, false);
+    return module !== undefined ? (module as JavaScriptModule) : undefined;
+  }
+
+  /**
+   * Load the specified WebAssembly command module.
+   * If loading fails return undefined, and caller is responsible for reporting this to the user if
+   * appropriate.
+   */
+  public getWasmModule(
     packageName: string,
     moduleName: string
   ): typeof EmscriptenModuleFactory | undefined {
+    const module = this._getModule(packageName, moduleName, true);
+    return module !== undefined ? (module as typeof EmscriptenModuleFactory) : undefined;
+  }
+
+  private _getModule(packageName: string, moduleName: string, wasm: boolean): any {
     let module = this.cache.get(packageName, moduleName);
     if (module === undefined) {
       // Maybe should use @jupyterlab/coreutils.URLExt to combine URL components.
       const filename = this.cache.key(packageName, moduleName) + '.js';
       const url = this.wasmBaseUrl + filename;
-      console.log('Cockle importing JavaScript/WebAssembly from ' + url);
+      console.log(`Cockle loading ${wasm ? 'WebAssembly' : 'JavaScript'} module from ${url}`);
 
       this.downloadModuleCallback(packageName, moduleName, true);
       try {
         importScripts(url);
-        module = (self as any).Module as typeof EmscriptenModuleFactory;
+        module = (self as any).Module;
       } finally {
         this.downloadModuleCallback(packageName, moduleName, false);
       }
