+++ src/types/javascript_module.d.ts
@@ -1,0 +1,5 @@
+import { IContext } from '../context';
+
+export interface JavaScriptModule {
+  run(cmdName: string, context: IContext): Promise<number>;
+}
