+++ test/js-util/js-capitalise.js
@@ -1,0 +1,8 @@
+async function run(cmdName, context) {
+  context.stdout.write(cmdName + ': ' + context.args.map(arg => arg.toUpperCase()).join(','));
+  return 1;
+}
+
+var Module = (() => {
+  return { run };
+})();
