+++ test/js-util/js-test.js
@@ -1,0 +1,8 @@
+async function run(cmdName, context) {
+  context.stdout.write(cmdName + ': ' + context.args.join(','));
+  return 0;
+}
+
+var Module = (() => {
+  return { run };
+})();
